Refactor transcription streaming implementations for AssemblyAI, Deepgram, and ElevenLabs
